tests: rename fixture
