+++ __fixtures__/mv2-kitchen-sink-paths.ts
