+++ src/html-inputs/__tests__/html__hook--buildStart.test.ts
@@ -16,7 +16,7 @@ import {
   optionsTsx,
   popupHtml,
   popupJs,
-} from '../../../__fixtures__/kitchen-sink-paths'
+} from '../../../__fixtures__/mv2-kitchen-sink-paths'
 import { context } from '../../../__fixtures__/plugin-context'
 import { loadHtml } from '../cheerio'
 import htmlInputs from '..'
