+++ src/html-inputs/__tests__/html__unit--getCss.test.ts
@@ -2,7 +2,7 @@ import { loadHtml, getCssHrefs } from '../cheerio'
 import {
   kitchenSinkRoot,
   optionsHtml,
-} from '../../../__fixtures__/kitchen-sink-paths'
+} from '../../../__fixtures__/mv2-kitchen-sink-paths'
 
 const html$ = loadHtml(kitchenSinkRoot)(optionsHtml)
