+++ src/html-inputs/__tests__/html__unit--mapScriptElems.test.ts
@@ -1,7 +1,7 @@
 import {
   kitchenSinkRoot,
   optionsHtml,
-} from '../../../__fixtures__/kitchen-sink-paths'
+} from '../../../__fixtures__/mv2-kitchen-sink-paths'
 import {
   mutateScriptElems,
   loadHtml,
