+++ src/manifest-input/__tests__/manifest__hook--buildStart.test.ts
@@ -11,7 +11,7 @@ import {
   missaaliOtf,
   notoSansLight,
   notoSansBlack,
-} from '../../../__fixtures__/kitchen-sink-paths'
+} from '../../../__fixtures__/mv2-kitchen-sink-paths'
 import { context as minContext } from '../../../__fixtures__/minimal-plugin-context'
 import { context } from '../../../__fixtures__/plugin-context'
 import {
