+++ src/manifest-input/manifest-parser/__tests__/mv2-manifest__unit--deriveFiles.test.ts
@@ -12,7 +12,7 @@ import {
   optionsHtml,
   popupHtml,
   srcDir,
-} from '../../../../__fixtures__/kitchen-sink-paths'
+} from '../../../../__fixtures__/mv2-kitchen-sink-paths'
 
 const manifest = readJSONSync(manifestJson)
